Patch to include CI with Actions (#11)
Now automatic testing is launch on push. 

The testing is done with PyTest via Github Actions. More tests are needed to cover all the cpp module. 

The setup.py containt dependencies that are automatically when the cpp module is install. For now the module is install with the pip -e option. Without this option pytest can't find the module after installation